FEAT: 여행 계획 목록 조회 시 사용자 token 검사 기능 추가

+++ src/main/java/com/ssafy/plan/controller/PlanController.java
@@ -4,9 +4,12 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
+import javax.servlet.http.HttpServletRequest;
+
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -51,10 +54,22 @@ public class PlanController {
 		@ApiResponse(code= 200, message = "서버 에러가 발생했습니다."),
 	})
 	@GetMapping("/{userid}")
-	public ResponseEntity<List<PlanDto>> planList(@PathVariable("userid") String userId) throws Exception {
+	public ResponseEntity<List<PlanDto>> planList(@PathVariable("userid") String userId, HttpServletRequest request) throws Exception {
 		logger.info("PlanController findPlansByMemberId - userId: {}.", userId);
-		
-		return ResponseEntity.ok(planService.findPlansByUserId(userId));
+
+		if (jwtService.checkToken(request.getHeader("access-token"))) {
+			logger.info("사용 가능한 토큰");
+			
+			try {
+				return new ResponseEntity<List<PlanDto>>(planService.findPlansByUserId(userId), HttpStatus.ACCEPTED);
+			} catch (Exception e) {
+				logger.error("정보조회 실패 : {}", e);
+				return new ResponseEntity<List<PlanDto>> (HttpStatus.INTERNAL_SERVER_ERROR);
+			}
+		} else {
+			logger.info("사용 불가능한 토큰");
+			return new ResponseEntity<List<PlanDto>> (HttpStatus.UNAUTHORIZED);
+		}
 	}
 
 	@ApiOperation(value = "여행 계획 추가", notes = "여행 계획을 추가합니다.")
@@ -64,12 +79,23 @@ public class PlanController {
 		@ApiResponse(code= 200, message = "서버 에러가 발생했습니다."),
 	})
 	@PostMapping
-	public ResponseEntity<List<PlanDto>> add(@RequestBody PlanDto plan) throws Exception {
+	public ResponseEntity<Object> add(@RequestBody PlanDto plan, HttpServletRequest request) throws Exception {
 		logger.info("PlanController add - plan :{}.", plan);
 		
-		planService.addPlan(plan);
-		
-		return ResponseEntity.ok().build();
+		if (jwtService.checkToken(request.getHeader("access-token"))) {
+			logger.info("사용 가능한 토큰");
+			
+			try {
+				planService.addPlan(plan);
+				return new ResponseEntity<Object>(HttpStatus.ACCEPTED);
+			} catch (Exception e) {
+				logger.error("정보조회 실패 : {}", e);
+				return new ResponseEntity<Object> (HttpStatus.INTERNAL_SERVER_ERROR);
+			}
+		} else {
+			logger.info("사용 불가능한 토큰");
+			return new ResponseEntity<Object> (HttpStatus.UNAUTHORIZED);
+		}
 	}
 	
 	@ApiOperation(value = "여행 계획 삭제", notes = "여행 계획을 삭제합니다.")
@@ -79,13 +105,25 @@ public class PlanController {
 		@ApiResponse(code= 200, message = "서버 에러가 발생했습니다."),
 	})
 	@DeleteMapping("/{planId}")
-	public ResponseEntity<Object> delete(@PathVariable Long planId) throws Exception {
+	public ResponseEntity<Object> delete(@PathVariable Long planId, HttpServletRequest request) throws Exception {
 		logger.info("PlanController delete - planId : {}", planId);
 
-		locationService.deleteLocationsByPlanId(planId);		
-		planService.deletePlan(planId);
-		
-		return ResponseEntity.ok().build();
+		if (jwtService.checkToken(request.getHeader("access-token"))) {
+			logger.info("사용 가능한 토큰");
+			
+			try {
+				locationService.deleteLocationsByPlanId(planId);		
+				planService.deletePlan(planId);
+				
+				return new ResponseEntity<Object>(HttpStatus.ACCEPTED);
+			} catch (Exception e) {
+				logger.error("정보조회 실패 : {}", e);
+				return new ResponseEntity<Object> (HttpStatus.INTERNAL_SERVER_ERROR);
+			}
+		} else {
+			logger.info("사용 불가능한 토큰");
+			return new ResponseEntity<Object> (HttpStatus.UNAUTHORIZED);
+		}
 	}
 	
 	@ApiOperation(value = "여행 계획 상세 정보 조회", notes = "여행 계획의 상세 정보를 반환합니다.")
@@ -95,13 +133,25 @@ public class PlanController {
 		@ApiResponse(code= 200, message = "서버 에러가 발생했습니다."),
 	})
 	@GetMapping("/view/{planId}")
-	public ResponseEntity<PlanDto> view(@PathVariable Long planId) throws Exception {
+	public ResponseEntity<PlanDto> view(@PathVariable Long planId, HttpServletRequest request) throws Exception {
 		logger.info("PlanController view - planId : {}.", planId);
 		
-		PlanDto plan = planService.findPlanById(planId);
-		plan.setLocations(locationService.findLocationsByPlanId(planId));
-		
-		return ResponseEntity.ok(plan);
+		if (jwtService.checkToken(request.getHeader("access-token"))) {
+			logger.info("사용 가능한 토큰");
+			
+			try {
+				PlanDto plan = planService.findPlanById(planId);
+				plan.setLocations(locationService.findLocationsByPlanId(planId));
+				
+				return new ResponseEntity<PlanDto>(plan, HttpStatus.ACCEPTED);
+			} catch (Exception e) {
+				logger.error("정보조회 실패 : {}", e);
+				return new ResponseEntity<PlanDto> (HttpStatus.INTERNAL_SERVER_ERROR);
+			}
+		} else {
+			logger.info("사용 불가능한 토큰");
+			return new ResponseEntity<PlanDto> (HttpStatus.UNAUTHORIZED);
+		}
 	}
 	
 	@ApiOperation(value = "여행 계획 일별 정보 조회", notes = "여행 계획의 일별 상세 정보를 반환합니다.")
@@ -111,19 +161,31 @@ public class PlanController {
 		@ApiResponse(code= 200, message = "서버 에러가 발생했습니다."),
 	})
 	@GetMapping("/view/{planId}/{day}")
-	public ResponseEntity<List<LocationDto>> view(@PathVariable Long planId, @PathVariable Integer day) throws Exception {
+	public ResponseEntity<List<LocationDto>> view(@PathVariable Long planId, @PathVariable Integer day, HttpServletRequest request) throws Exception {
 		logger.info("PlanController view - planId : {}, day: {}.", planId, day);
 
-		Map<String, Object> map = new HashMap<String, Object>();
-		map.put("planId", planId);
-		map.put("day", day);
-		
-		List<LocationDto> locations = locationService.findLocationsByPlanIdAndDay(map);
-		for(LocationDto location : locations) {
-			location.setAttraction(attractionService.findAttractionByContentId(location.getContentId()));
+		if (jwtService.checkToken(request.getHeader("access-token"))) {
+			logger.info("사용 가능한 토큰");
+			
+			try {
+				Map<String, Object> map = new HashMap<String, Object>();
+				map.put("planId", planId);
+				map.put("day", day);
+				
+				List<LocationDto> locations = locationService.findLocationsByPlanIdAndDay(map);
+				for(LocationDto location : locations) {
+					location.setAttraction(attractionService.findAttractionByContentId(location.getContentId()));
+				}
+				
+				return new ResponseEntity<List<LocationDto>>(locations, HttpStatus.ACCEPTED);
+			} catch (Exception e) {
+				logger.error("정보조회 실패 : {}", e);
+				return new ResponseEntity<List<LocationDto>> (HttpStatus.INTERNAL_SERVER_ERROR);
+			}
+		} else {
+			logger.info("사용 불가능한 토큰");
+			return new ResponseEntity<List<LocationDto>> (HttpStatus.UNAUTHORIZED);
 		}
-		
-		return ResponseEntity.ok(locations);
 	}
 	
 	@ApiOperation(value = "여행 계획 수정", notes = "여행 계획을 수정합니다.")
@@ -133,15 +195,27 @@ public class PlanController {
 		@ApiResponse(code= 200, message = "서버 에러가 발생했습니다."),
 	})
 	@PutMapping
-	public ResponseEntity<PlanDto> update(@RequestBody PlanDto plan) throws Exception {
+	public ResponseEntity<PlanDto> update(@RequestBody PlanDto plan, HttpServletRequest request) throws Exception {
 		logger.info("PlanController update - plan : {}.", plan);
 		
-		planService.updatePlan(plan);
-		for (LocationDto location : plan.getLocations()) {
-			locationService.updateLocation(location);
+		if (jwtService.checkToken(request.getHeader("access-token"))) {
+			logger.info("사용 가능한 토큰");
+			
+			try {
+				planService.updatePlan(plan);
+				for (LocationDto location : plan.getLocations()) {
+					locationService.updateLocation(location);
+				}
+				
+				return new ResponseEntity<PlanDto>(planService.findPlanById(plan.getId()), HttpStatus.ACCEPTED);
+			} catch (Exception e) {
+				logger.error("정보조회 실패 : {}", e);
+				return new ResponseEntity<PlanDto> (HttpStatus.INTERNAL_SERVER_ERROR);
+			}
+		} else {
+			logger.info("사용 불가능한 토큰");
+			return new ResponseEntity<PlanDto> (HttpStatus.UNAUTHORIZED);
 		}
-		
-		return ResponseEntity.ok(planService.findPlanById(plan.getId()));
 	}
 	
 	@ApiOperation(value = "여행 계획에 장소 추가", notes = "여행 계획에 장소를 추가합니다.")
@@ -151,12 +225,24 @@ public class PlanController {
 		@ApiResponse(code= 200, message = "서버 에러가 발생했습니다."),
 	})
 	@PostMapping("/location")
-	public ResponseEntity<Object> addLocation(@RequestBody LocationDto location) throws Exception {
+	public ResponseEntity<Object> addLocation(@RequestBody LocationDto location, HttpServletRequest request) throws Exception {
 		logger.info("PlanController add Location - location : {}.", location);
-
-		locationService.addLocation(location);
 		
-		return ResponseEntity.ok().build();
+		if (jwtService.checkToken(request.getHeader("access-token"))) {
+			logger.info("사용 가능한 토큰");
+			
+			try {
+				locationService.addLocation(location);
+				
+				return new ResponseEntity<Object>(HttpStatus.ACCEPTED);
+			} catch (Exception e) {
+				logger.error("정보조회 실패 : {}", e);
+				return new ResponseEntity<Object> (HttpStatus.INTERNAL_SERVER_ERROR);
+			}
+		} else {
+			logger.info("사용 불가능한 토큰");
+			return new ResponseEntity<Object> (HttpStatus.UNAUTHORIZED);
+		}
 	}
 	
 	@ApiOperation(value = "장소 삭제", notes = "여행 계획에서 장소를 삭제합니다.")
@@ -166,11 +252,23 @@ public class PlanController {
 		@ApiResponse(code= 200, message = "서버 에러가 발생했습니다."),
 	})
 	@DeleteMapping("/location/{locationId}")
-	public ResponseEntity<Object> deleteLocation(@PathVariable Long locationId) throws Exception {
+	public ResponseEntity<Object> deleteLocation(@PathVariable Long locationId, HttpServletRequest request) throws Exception {
 		logger.info("PlanController delete Location - locationId : {}.", locationId);
 
-		locationService.deleteLocation(locationId);
-
-		return ResponseEntity.ok().build();
+		if (jwtService.checkToken(request.getHeader("access-token"))) {
+			logger.info("사용 가능한 토큰");
+			
+			try {
+				locationService.deleteLocation(locationId);
+				
+				return new ResponseEntity<Object>(HttpStatus.ACCEPTED);
+			} catch (Exception e) {
+				logger.error("정보조회 실패 : {}", e);
+				return new ResponseEntity<Object> (HttpStatus.INTERNAL_SERVER_ERROR);
+			}
+		} else {
+			logger.info("사용 불가능한 토큰");
+			return new ResponseEntity<Object> (HttpStatus.UNAUTHORIZED);
+		}
 	}
 }
